Split tree interreter from jvm interpreter
